% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chile_candidate_votes_2021}
\alias{chile_candidate_votes_2021}
\title{Chilean 2021 Presidential Election Votes}
\format{
A data frame with 46,639 rows and 14 variables:
\describe{
\item{\code{REGION}}{Geographic region of the ballot box.}
\item{\code{CIRCUNSCRIPCION.ELECTORAL}}{The electoral circumscription of the ballot box.}
\item{\code{LOCAL}}{The name of the location where the voting took place (usually a school).}
\item{\code{MESA}}{An identifier for the ballot box within the \code{LOCAL}.}
\item{\code{EDUARDO.ARTES}}{The number of votes cast for the candidate \emph{Eduardo Artés}.}
\item{\code{FRANCO.ALDO.PARISI}}{The number of votes cast for the candidate \emph{Franco Parisi}.}
\item{\code{GABRIEL.BORIC}}{The number of votes cast for the candidate \emph{Gabriel Boric}.}
\item{\code{JOSE.ANTONIO.KAST}}{The number of votes cast for the candidate \emph{José Antonio Kast}.}
\item{\code{MARCO.ENRIQUEZ.OMINAMI}}{The number of votes cast for the candidate \emph{Marco Enríquez-Ominami}.}
\item{\code{SEBASTIA.SICHEL}}{The number of votes cast for the candidate \emph{Sebastián Sichel}.}
\item{\code{VOTOS.EN.BLANCO}}{The number of blank votes.}
\item{\code{VOTOS.NULO}}{The number of null votes.}
\item{\code{YASNA.PROVOSTE}}{The number of votes cast for the candidate \emph{Yasna Provoste}.}
\item{\code{NULO.BLANCO}}{The sum of \code{VOTOS.NULO} and \code{VOTOS.EN.BLANCO}.}
}
}
\source{
\href{https://www.servel.cl/}{Chilean Electoral Service (SERVEL)}
}
\usage{
data("chile_candidate_votes_2021")
}
\description{
This dataset contains the results of the first round of the 2021 Chilean presidential elections.
It provides 10 possible voting options. Each ballot box is identified by a ballot box number
(\code{MESA}), voting location (\code{LOCAL}), electoral circumscription
(\code{CIRCUNSCRIPCION.ELECTORAL}), and region (\code{REGION}).
}
\examples{
data("chile_candidate_votes_2021")
head(chile_candidate_votes_2021)

}
\seealso{
chile_groups_2021
}
\keyword{datasets}
