% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eimodel-package.R
\docType{package}
\name{eimodel-package}
\alias{eimodel}
\alias{eimodel-package}
\title{eimodel: A Fast Alternative for the R x C Ecological Inference Problem}
\description{
Addressing the \eqn{R \times C} ecological inference problem is challenging due to
the variety of possible approaches, including parametric probability models,
entropy maximization, or mathematical programming. This package introduces a
nonparametric model, based on the Expectation-Maximization (EM) algorithm, to maximize
the likelihood given observed data. Although the \emph{M}-step can be solved
in closed form, the \emph{E}-step can involve an exponential number of computations.
}
\details{
The \code{eimodel} package provides multiple methods to estimate the \emph{E}-step
in an EM algorithm for the standard \eqn{R \times C} ecological inference problem.
It also offers a function to simulate elections under diverse parameters. These
methods, which can be accessed via the \code{eim} object, include:
\itemize{
\item \verb{Hit and Run (H&R)}: Simulates a subset of samples to approximate the exact solution.
\item \verb{Multivariate Normal PDF (MVN PDF)}: Approximates the \emph{E}-step using a multivariate normal PDF.
\item \verb{Multivariate Normal CDF (MVN CDF)}: Uses the CDF of a multivariate normal distribution to approximate probabilities.
\item \code{Multinomial (Mult)}: Approximates the \emph{E}-step with a multinomial distribution.
\item \code{Exact}: Solves the determinant approach directly.
}

On average, the \code{Multinomial} method tends to be the most efficient and precise,
often yielding the lowest mean absolute error (MAE). Its runtime
is roughly on the order of \eqn{10^{-5}} seconds.

To learn more about \code{eimodel}, please consult the available vignettes:

\preformatted{
browseVignettes("eimodel")
}
}
\references{
\href{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=4832834}{Thraves, C. and Ubilla, P.: \emph{"Fast Ecological Inference Algorithm for the R×C Case"}.}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/DanielHermosilla/ecological-inference-elections/issues}
}

}
