% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eim-package.R
\name{sum.eim}
\alias{sum.eim}
\title{Returns the total amount of voters in the system.}
\usage{
\method{sum}{eim}(object)
}
\arguments{
\item{object}{An eim object.}
}
\value{
The amount of voters.
}
\description{
Given a initialized object, it returns the total amount of voters. It's equivalent
of running \code{sum(object$W)} or \code{sum(object$X)}.
}
\examples{
simulations <- simulate_votation(
  num_ballots = 20,
  num_candidates = 5,
  num_demographics = 3,
  ballot_voters = rep(100, 20)
)

model <- eim$new(simulations$X, simulations$W)
sum(model) # 2000

}
