% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eim-package.R
\name{update.eim}
\alias{update.eim}
\title{Update an existing eim model with a new EM algorithm computation}
\usage{
\method{update}{eim}(object, ...)
}
\arguments{
\item{...}{New parameters to passed to \code{compute()}.}

\item{\emph{object}}{An eim object.}
}
\value{
Changes made on the updated object.
}
\description{
This function updates an object with a new Expected Maximization computation
with other parameters.
}
\examples{
simulations <- simulate_votation(
  num_ballots = 20,
  num_candidates = 5,
  num_demographics = 3,
  ballot_voters = rep(100, 20)
)

model <- eim$new(simulations$X, simulations$W)

predict(model)

update(model, "MVN PDF")

model$method # Returns 'MVN PDF'

}
