% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eim-class.R
\name{bootstrap}
\alias{bootstrap}
\title{Runs a Bootstrap to Estimate the \strong{Standard Deviation} of Predicted Probabilities}
\usage{
bootstrap(
 boot_iter,
 object = NULL,
 X = NULL,
 W = NULL,
 json_path = NULL,
 mismatch = FALSE,
	get_p = TRUE,
 ...
)
}
\arguments{
\item{boot_iter}{Integer specifying how many times to run the
EM algorithm.}

\item{object}{An object of class \code{eim}, which can be created using the \link{eim} function.
This parameter shouldn't be called if you prefer to supply the \code{X} and \code{W} matrices or a \code{json_path}.}

\item{X}{A \verb{(b x c)} matrix representing candidate votes per ballot box.}

\item{W}{A \verb{(b x g)} matrix representing group votes per ballot box.}

\item{json_path}{A path to a JSON file containing \code{X} and \code{W} fields, stored as nested arrays.}

\item{get_p}{Boolean that determines if the probability is estimated aswell. Defaults to \code{TRUE}.}

\item{...}{Additional arguments passed to the \link{compute} function that will execute the EM algorithm.}
}
\value{
Updates the \code{eim} object with a new \code{sd} field containing the estimated
standard deviations of the probabilities.
}
\description{
This function computes the Expected-Maximization (EM) algorithm "\code{boot_iter}" times by sampling,
with replacement, all of the ballot boxes out of the original data for each
iteration. A new sample is drawn for every iteration, and from all of these iterations,
the \strong{standard deviation} of the estimated probabilities is computed.
}
\note{
This function can be executed using one of three mutually exclusive approaches:
\enumerate{
\item By providing an existing \code{eim} object.
\item By supplying both input matrices (\code{X} and \code{W}) directly.
\item By specifying a JSON file (\code{json_path}) containing the matrices.
}

These input methods are \strong{mutually exclusive}, meaning that you must provide \strong{exactly one} of
these options. Attempting to provide more than one or none of these inputs will result in an error.

When called with an \code{eim} object, the function updates the object with the computed results.
If an \code{eim} object is not provided, the function will create one internally using either the
supplied matrices or the data from the JSON file before executing the algorithm.
}
\examples{
# Example 1: Using an 'eim' object directly
simulations <- simulate_election(
    num_ballots = 20,
    num_candidates = 5,
    num_groups = 3,
    ballot_voters = rep(100, 20)
)

# Initialize the 'eim' object
model <- eim(X = simulations$X, W = simulations$W)

# Run bootstrap with an existing 'eim' object
model <- bootstrap(
    boot_iter = 30,
    object = model,
    method = "mvn_cdf",
    maxiter = 500,
    verbose = TRUE,
    mc_samples = 5000
)

# Access standard deviation and probability results
print(model$sd)
print(model$prob)


# Example 2: Providing 'X' and 'W' matrices directly
model <- bootstrap(
    boot_iter = 50,
    get_p = FALSE,
    X = simulations$X,
    W = simulations$W,
    method = "exact",
    maxiter = 100,
    maxtime = 15,
    stop_threshold = 0.01
)

print(model$sd)

# Example 3: Using a JSON file as input
\dontrun{
model <- bootstrap(
    boot_iter = 20,
    json_path = "path/to/election_data.json",
    method = "mult",
    get_p = FALSE
)

print(model$sd)
}

}
\seealso{
The \link{eim} object and \link{compute} implementation.
}
